Add CSRF protection to deleting photos

+++ fotoboek.php
@@ -574,9 +574,12 @@
 			if (!$this->policy->user_can_update($book))
 				throw new UnauthorizedException();
 			
-			foreach ($_POST['photo'] as $id)
-				if ($photo = $this->model->get_iter($id))
-					$this->model->delete($photo);
+			if ($this->_form_is_submitted('delete_photos'))
+			{
+				foreach ($_POST['photo'] as $id)
+					if ($photo = $this->model->get_iter($id))
+						$this->model->delete($photo);
+			}
 			
 			return $this->view->redirect('fotoboek.php?book=' . $book->get_id());
 		}
